src/requirements.txt
streamlit
psycopg2-binary
pandas
plotly
python-dotenv
pydantic
faker